+++ secrets/secrets-sdk/components.ts
@@ -2,4 +2,4 @@
 export { default as SecretsTable } from "./displays/secrets-table/SecretsTable.vue";
 export { default as PendingApprovalsTable } from "./displays/pending-approvals-table/PendingApprovalsTable.vue";
 export { default as MissingSecretsTable } from "./displays/missing-secrets-table/MissingSecretsTable.vue";
-export { default as SecretForm } from "./forms/secret-form/SecretForm.vue";
+export { default as CreateSecretForm } from "./forms/secret-form/CreateSecretForm.vue";
