+++ secrets/secrets-sdk/strings.ts
@@ -1,6 +1,6 @@
 import { secrets_table_strings } from "./displays/secrets-table/SecretsTable.strings.ts";
 import { pending_approvals_table_strings } from "./displays/pending-approvals-table/PendingApprovalsTable.strings.ts";
-import { secret_form_strings } from "./forms/secret-form/SecretForm.strings.ts";
+import { secret_form_strings } from "./forms/secret-form/CreateSecretForm.strings.ts";
 
 export const secretsSdkStrings = {
   secrets_table_strings,
